# auth-service/src/main/java/com/vet/auth_service/api/dto/AppointmentDto.java
package com.vet.auth_service.api.dto;

import java.time.LocalDateTime;

public record AppointmentDto(
        String id,
        LocalDateTime dateTime,
        String petName,
        String ownerName,
        String status
) {}