# auth-service/src/main/java/com/vet/auth_service/api/dto/UserRegistrationDto.java
package com.vet.auth_service.api.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Builder;

@Builder
public record UserRegistrationDto(
        @NotBlank String userId,
        @NotBlank String firstName,
        @NotBlank String lastName,
        String email
) {
}