# profile-service/src/main/java/com/vet/profile_service/api/dto/AppointmentDto.java
package com.vet.profile_service.api.dto;

import java.time.LocalDateTime;

public record AppointmentDto(
        String id,
        LocalDateTime dateTime,
        String petName,
        String ownerName,
        String status
) {}